adjust css on cart.css
